Updated HUD Icons
As per the user story C4P-12, I have updated the uasset files associated with HUD Icons to be a bit more fleshed out and more in line with the original unreal tournament. Something that should be noted is that some of the icons colors do not match perfectly so perhaps that is something to be tweaked later if necessary.