MetaSound Implementation - Terminal Ambience
Implemented the first MetaSound (as far as I could see), which is the upgraded version of UE4's "Sound Cue" for UE5. More details follow, but first:
MetaSound can be found in the Audio folder: "MS_TerminalChime"
The MetaSound was added to the map "Aqua_2" and can be found at the three screens/monitors/terminals in the map.

MetaSounds, as this example will show, are the best thing to use moving forward. Here are some of the advantages:
- Reducing overall game size.
- Allowing for even more in-engine digital signal processing, 'live' mixing etc.
- Creating even more robust sound systems / hierarchies than before.

MS_TerminalChime is a simple example / test to show the usefulness of MetaSounds. The implementation is simple: A loop of a chime I made that plays when nearing aforementioned terminals in maps - an object ambience.

Going into MS_TerminalChime you can see the only two functions are a wave player (basically required) and a random(f). The random is where the magic happens.

I have randomized the pitch of the sound file to a range of 6 semitones 'surrounding' the original. Every time someone enters the attenuation volume of the MetaSound, the sound file will be played at a different pitch. If you want to see it work more clearly, increase the min/max pitch range in the MetaSound flow graph and give it a spin.

The result:
- Programmed variety that adds more to the general ambience.
- Only using a single, short, small-size .wav file to generate an effect that might otherwise take several files.

The last part is perhaps the most important. We can now apply this process to more tedious things such as footsteps! Instead of recording 4+ different variations of the same step - bloating the game's size in the process - we can use a single step sound file and apply various effects to it, in-engine, to create variation.

Furthermore, we can even apply separate in-engine effects, depending on the surface (steel, gravel, dirt) or the environment (land, underwater), to change the sound of the step on a whim as long as the categorization structure of different materials or volumes are appropriately set.

This process can help streamline a lot of different parts of the soundscape to work more closely together while taking up less overall space.